sent test function to module file
